--- BlogProject/BlogApp/templates/blog/post_list.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Post List</title>
-</head>
-<body>
-    <header>
-        <nav>
-            <a href="{% url 'post_list' %}">Posts</a>
-            <a href="{% url 'create_post' %}">Create Post</a>
-            <!-- Other links -->
-        </nav>
-    </header>
-    
-    <main>
-        <h1>All Posts</h1>
-        <ul>
-            {% for post in posts %}
-                <li><a href="{% url 'post_detail' post.id %}">{{ post.title }}</a></li>
-            {% endfor %}
-        </ul>
-    </main>
-</body>
-</html>
