+++ BlogProject/BlogApp/views.py
@@ -1,11 +1,17 @@
-from django.http import HttpResponseBadRequest
+from django.http import HttpResponse, HttpResponseBadRequest
 from django.shortcuts import render, redirect
 from django.contrib.auth import login as auth_login, authenticate
 from django.contrib.auth import logout as auth_logout
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import get_object_or_404
 
-from .models import CommentAttachment, User, Post, Comment, Report, Suggestion
+from suggestions.models import Suggestion, SuggestionReply
+
+from .models import User
+from reports.models import Report
+from comments.models import Comment,CommentAttachment
+from posts.models import Post
+from suggestions.models import Suggestion, SuggestionReply
 from .forms import CommentAttachmentForm, CommentForm, ReportForm, SuggestionForm
 from .forms import CustomUserCreationForm, CustomAuthenticationForm
 from .forms import CustomUserCreationForm
@@ -91,182 +97,190 @@ def logout(request):
 @login_required
 def dashboard(request):
     user = request.user  # Get the current logged-in user
+
+    # Fetch recent comments on posts created by the user
     recent_comments_on_posts = Comment.objects.filter(
         post__in=user.posts.all()
     ).order_by('-created_at')[:5]
 
-
     # Count the total posts made by the user
     posts_count = user.posts.count()
 
-    # Get the most recent post made by the user
+    # Get the most recent posts made by the user
     recent_posts = user.posts.order_by('-created_at')[:5]  # Get the 5 most recent posts
 
-    # Get the most recent likes made by the user from the post table as there is no like table
-     # Fetch recent likes for the posts
-
+    # Get the most recent likes made by the user from the post table (assuming a 'likes' field exists)
     posts = Post.objects.prefetch_related('likes').all()
-    # Create a list of posts with their likes
+
+    # Check if the user wants to see suggestions
+    show_suggestions = request.GET.get('section') == 'suggestions'
+
+    # Fetch suggestions if the user wants to see them
+    suggestions = []
+    if show_suggestions:
+        suggestions = Suggestion.objects.filter(post__author=user)
 
     context = {
         'user': user,
         'recent_comments_on_posts': recent_comments_on_posts,
         'posts_count': posts_count,
         'recent_posts': recent_posts,
-        'posts': posts
+        'posts': posts,
+        'suggestions': suggestions,
+        'show_suggestions': show_suggestions,
     }
+
     return render(request, 'blog/dashboard.html', context)
 
+# @login_required
+# def create_post(request):
+#     if request.method == 'POST':
+#         form = PostForm(request.POST, request.FILES)
+#         if form.is_valid():
+#             post = form.save(commit=False)
+#             post.author = request.user
+#             post.save()
 
-@login_required
-def create_post(request):
-    if request.method == 'POST':
-        form = PostForm(request.POST, request.FILES)
-        if form.is_valid():
-            post = form.save(commit=False)
-            post.author = request.user
-            post.save()
+#             # Handle single file upload
+#             file = request.FILES.get('attachment')
+#             form.save_attachment(post, file)
 
-            # Handle single file upload
-            file = request.FILES.get('attachment')
-            form.save_attachment(post, file)
+#             return redirect('dashboard')  # Adjust as needed
+#     else:
+#         form = PostForm()
 
-            return redirect('dashboard')  # Adjust as needed
-    else:
-        form = PostForm()
+#     return render(request, 'blog/create_post.html', {'form': form})
 
-    return render(request, 'blog/create_post.html', {'form': form})
+# @login_required
+# def post_detail(request, post_id):
+#     post = get_object_or_404(Post, id=post_id)
+#     comments = post.comments.filter(parent_comment__isnull=True)
 
-@login_required
-def post_detail(request, post_id):
-    post = get_object_or_404(Post, id=post_id)
-    comments = post.comments.filter(parent_comment__isnull=True)
+#     if request.method == 'POST':
+#         # Handle comment submission with attachments
+#         if 'submit_comment' in request.POST:
+#             comment_form = CommentForm(request.POST)
+#             if comment_form.is_valid():
+#                 comment = comment_form.save(commit=False)
+#                 comment.post = post
+#                 comment.author = request.user
+#                 comment.save()
 
-    if request.method == 'POST':
-        # Handle comment submission with attachments
-        if 'submit_comment' in request.POST:
-            comment_form = CommentForm(request.POST)
-            if comment_form.is_valid():
-                comment = comment_form.save(commit=False)
-                comment.post = post
-                comment.author = request.user
-                comment.save()
+#                 # Handle attachments
+#                 attachment_form = CommentAttachmentForm(request.POST, request.FILES)
+#                 if attachment_form.is_valid():
+#                     for attachment in request.FILES.getlist('file'):
+#                         CommentAttachment.objects.create(comment=comment, file=attachment)
 
-                # Handle attachments
-                attachment_form = CommentAttachmentForm(request.POST, request.FILES)
-                if attachment_form.is_valid():
-                    for attachment in request.FILES.getlist('file'):
-                        CommentAttachment.objects.create(comment=comment, file=attachment)
+#                 return redirect('post_detail', post_id=post.id)
 
-                return redirect('post_detail', post_id=post.id)
+#         # Handle report submission
+#         if 'submit_report' in request.POST:
+#             report_form = ReportForm(request.POST)
+#             if report_form.is_valid():
+#                 report = report_form.save(commit=False)
+#                 report.post = post
+#                 report.reported_by = request.user
+#                 report.save()
+#                 return redirect('post_detail', post_id=post.id)
 
-        # Handle report submission
-        if 'submit_report' in request.POST:
-            report_form = ReportForm(request.POST)
-            if report_form.is_valid():
-                report = report_form.save(commit=False)
-                report.post = post
-                report.reported_by = request.user
-                report.save()
-                return redirect('post_detail', post_id=post.id)
+#         # Handle suggestion submission
+#         if 'submit_suggestion' in request.POST:
+#             suggestion_form = SuggestionForm(request.POST)
+#             if suggestion_form.is_valid():
+#                 suggestion = suggestion_form.save(commit=False)
+#                 suggestion.post = post
+#                 suggestion.user = request.user
+#                 suggestion.save()
+#                 return redirect('post_detail', post_id=post.id)
 
-        # Handle suggestion submission
-        if 'submit_suggestion' in request.POST:
-            suggestion_form = SuggestionForm(request.POST)
-            if suggestion_form.is_valid():
-                suggestion = suggestion_form.save(commit=False)
-                suggestion.post = post
-                suggestion.user = request.user
-                suggestion.save()
-                return redirect('post_detail', post_id=post.id)
+#     else:
+#         comment_form = CommentForm()
+#         attachment_form = CommentAttachmentForm()  # Initialize an empty form for attachments
+#         report_form = ReportForm()
+#         suggestion_form = SuggestionForm()
 
-    else:
-        comment_form = CommentForm()
-        attachment_form = CommentAttachmentForm()  # Initialize an empty form for attachments
-        report_form = ReportForm()
-        suggestion_form = SuggestionForm()
+#     context = {
+#         'post': post,
+#         'comments': comments,
+#         'comment_form': comment_form,
+#         'attachment_form': attachment_form,
+#         'report_form': report_form,
+#         'suggestion_form': suggestion_form,
+#     }
 
-    context = {
-        'post': post,
-        'comments': comments,
-        'comment_form': comment_form,
-        'attachment_form': attachment_form,
-        'report_form': report_form,
-        'suggestion_form': suggestion_form,
-    }
+#     return render(request, 'blog/post_detail.html', context)
 
-    return render(request, 'blog/post_detail.html', context)
+# @login_required
+# def like_post(request, post_id):
+#     post = get_object_or_404(Post, id=post_id)
+#     if request.user in post.likes.all():
+#         post.likes.remove(request.user)
+#     else:
+#         post.likes.add(request.user)
+#     return redirect('post_detail', post_id=post.id)
 
-@login_required
-def like_post(request, post_id):
-    post = get_object_or_404(Post, id=post_id)
-    if request.user in post.likes.all():
-        post.likes.remove(request.user)
-    else:
-        post.likes.add(request.user)
-    return redirect('post_detail', post_id=post.id)
+# @login_required
+# def like_comment(request, comment_id):
+#     comment = get_object_or_404(Comment, id=comment_id)
+#     if request.user in comment.likes.all():
+#         comment.likes.remove(request.user)
+#     else:
+#         comment.likes.add(request.user)
+#     return redirect('posts:post_detail', post_id=comment.post.id)
 
-@login_required
-def like_comment(request, comment_id):
-    comment = get_object_or_404(Comment, id=comment_id)
-    if request.user in comment.likes.all():
-        comment.likes.remove(request.user)
-    else:
-        comment.likes.add(request.user)
-    return redirect('post_detail', post_id=comment.post.id)
-
-@login_required
-def reply_to_comment(request, comment_id):
-    parent_comment = get_object_or_404(Comment, id=comment_id)
-    if request.method == 'POST':
-        form = CommentForm(request.POST)
-        if form.is_valid():
-            reply = form.save(commit=False)
-            reply.post = parent_comment.post
-            reply.author = request.user
-            reply.parent_comment = parent_comment
-            reply.save()
-            return redirect('post_detail', post_id=parent_comment.post.id)
-    else:
-        form = CommentForm()
+# @login_required
+# def reply_to_comment(request, comment_id):
+#     parent_comment = get_object_or_404(Comment, id=comment_id)
+#     if request.method == 'POST':
+#         form = CommentForm(request.POST)
+#         if form.is_valid():
+#             reply = form.save(commit=False)
+#             reply.post = parent_comment.post
+#             reply.author = request.user
+#             reply.parent_comment = parent_comment
+#             reply.save()
+#             return redirect('posts:post_detail', post_id=parent_comment.post.id)
+#     else:
+#         form = CommentForm()
     
-    return render(request, 'blog/reply_to_comment.html', {'form': form, 'parent_comment': parent_comment})
+#     return render(request, 'blog/reply_to_comment.html', {'form': form, 'parent_comment': parent_comment})
 
-@login_required
-def add_comment(request, post_id):
-    if request.method == 'POST':
-        form = CommentForm(request.POST, request.FILES)
-        if form.is_valid():
-            comment = form.save(commit=False)
-            comment.post_id = post_id
-            comment.author = request.user
-            comment.save()
+# @login_required
+# def add_comment(request, post_id):
+#     if request.method == 'POST':
+#         form = CommentForm(request.POST, request.FILES)
+#         if form.is_valid():
+#             comment = form.save(commit=False)
+#             comment.post_id = post_id
+#             comment.author = request.user
+#             comment.save()
 
-            # Handle attachments
-            attachments = request.FILES.getlist('attachments')
-            for attachment in attachments:
-                CommentAttachment.objects.create(comment=comment, file=attachment)
+#             # Handle attachments
+#             attachments = request.FILES.getlist('attachments')
+#             for attachment in attachments:
+#                 CommentAttachment.objects.create(comment=comment, file=attachment)
 
-            return redirect('post_detail', post_id=post_id)
-    else:
-        form = CommentForm()
+#             return redirect('post_detail', post_id=post_id)
+#     else:
+#         form = CommentForm()
 
-    return render(request, 'add_comment.html', {'form': form})
+#     return render(request, 'add_comment.html', {'form': form})
 
-@login_required
-def report_post(request, post_id):
-    if request.method == 'POST':
-        post = get_object_or_404(Post, id=post_id)
-        reason = request.POST.get('reason')
+# @login_required
+# def report_post(request, post_id):
+#     if request.method == 'POST':
+#         post = get_object_or_404(Post, id=post_id)
+#         reason = request.POST.get('reason')
         
-        Report.objects.create(
-            report_type='post',
-            post=post,
-            reported_by=request.user,
-            reason=reason
-        )
+#         Report.objects.create(
+#             report_type='post',
+#             post=post,
+#             reported_by=request.user,
+#             reason=reason
+#         )
         
-        return redirect('post_detail', post_id=post_id)
+#         return redirect('post_detail', post_id=post_id)
 
 @login_required
 def report_comment(request, comment_id):
@@ -288,36 +302,83 @@ def report_comment(request, comment_id):
             return redirect(referer)
         
         # Fallback redirect if referer is not available
-        return redirect('blog/post_detail')  # Change 'home' to a valid URL or view name
+        return redirect('posts/post_detail')  # Change 'home' to a valid URL or view name
     
     return HttpResponseBadRequest("Invalid request method.")
 
-@login_required
-def suggest_post(request, post_id):
-    if request.method == 'POST':
-        post = get_object_or_404(Post, id=post_id)
-        content = request.POST.get('content')
-        
-        # Simplified debugging
-        print("Request POST data:", request.POST)
-        print("Content from request.POST:", content)
 
-        if content:
-            Suggestion.objects.create(
-                post=post,
-                user=request.user,
-                content=content
+@login_required
+def reply_to_suggestion(request, suggestion_id):
+    suggestion = get_object_or_404(Suggestion, id=suggestion_id)
+    
+    if request.method == 'POST':
+        reply_content = request.POST.get('reply_content')
+        
+        if reply_content:
+            # Create a reply to the suggestion
+            reply = SuggestionReply.objects.create(
+                suggestion=suggestion,
+                user=request.user,  # Assuming this is the post author
+                content=reply_content
             )
+            print("Suggestion reply created:", reply)
+            
+            # Associate the reply with the suggestion
+            suggestion.reply = reply
+            suggestion.save()
+            
             return redirect(request.META.get('HTTP_REFERER', 'home'))
         else:
-            print("No content provided")
+            print("No reply content provided")
+            # Optionally, add a message to the user indicating that the reply content is required
+            # messages.error(request, "Content is required to submit a reply.")
     
- 
-    return redirect('blog/post_detail')  # Fallback redirect if the request method is not POST
+    return redirect('posts:post_detail', post_id=suggestion.post.id)  # Redirect to the post detail view
+
+# @login_required
+# def suggest_post(request, post_id):
+#     post = get_object_or_404(Post, id=post_id)
+    
+#     if request.method == 'POST':
+#         content = request.POST.get('content')
+        
+#         # Debugging output (optional)
+#         print("Request POST data:", request.POST)
+#         print("Content from request.POST:", content)
+
+#         if content:
+#             # Create a new suggestion
+#             suggestion = Suggestion.objects.create(
+#                 post=post,
+#                 user=request.user,
+#                 content=content
+#             )
+#             print("Suggestion created:", suggestion)
+            
+#             return redirect(request.META.get('HTTP_REFERER', 'home'))  # Redirect back to the previous page
+#         else:
+#             # Handle the case where no content was provided
+#             print("No content provided")
+#             # Optionally, add a message to the user indicating that the content is required
+#             # messages.error(request, "Content is required to submit a suggestion.")
+    
+#     return redirect('blog:post_detail', post_id=post.id)  # Redirect to the post detail view if not POST
+
+
+
+
+
+# @login_required
+# def post_list(request):
+#     posts = Post.objects.all()  # Fetch all posts
+#     return render(request, 'blog/post_list.html', {'posts': posts})
 
 @login_required
-def post_list(request):
-    posts = Post.objects.all()  # Fetch all posts
-    return render(request, 'blog/post_list.html', {'posts': posts})
+def jump_to_suggestions(request):
+    print("Jump to suggestions")
+    return redirect('suggestions:user_suggestions')
 
-
+@login_required
+def jump_to_posts(request):
+    print("Jump to posts")
+    return redirect('posts:post_list')
