+++ BlogProject/comments/models.py
@@ -1,0 +1,21 @@
+from django.db import models
+from django.contrib.auth import get_user_model
+
+User = get_user_model()  # Import User model if it's a custom model
+
+# Create your models here.
+# Comment model
+class Comment(models.Model):
+    post = models.ForeignKey('posts.Post', on_delete=models.CASCADE, related_name='comments')
+    author = models.ForeignKey('BlogApp.User', on_delete=models.CASCADE, related_name='comments')
+  
+    content = models.TextField()
+    created_at = models.DateTimeField(auto_now_add=True)
+    likes = models.ManyToManyField(User, related_name='liked_comments', blank=True)
+    parent_comment = models.ForeignKey('self', on_delete=models.CASCADE, related_name='replies', null=True, blank=True)
+
+# CommentAttachment model
+class CommentAttachment(models.Model):
+    comment = models.ForeignKey(Comment, on_delete=models.CASCADE, related_name='attachments')
+    file = models.FileField(upload_to='comment_attachments/')
+    uploaded_at = models.DateTimeField(auto_now_add=True)
