+++ BlogProject/comments/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+from . import views
+
+app_name = 'comments'
+
+urlpatterns = [
+    path('comment/<int:comment_id>/like/', views.like_comment, name='like_comment'),
+    path('comment/<int:comment_id>/reply/', views.reply_to_comment, name='reply_to_comment'),
+]
